WIP. Show example queries, no styling
