+++ src/shared/components/query/filteredSearch/FilteredSearchDropdownForm.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { FunctionComponent } from 'react';
-import { CancerTreeSearchFilter } from 'shared/lib/textQueryUtils';
+import { CancerTreeSearchFilter } from 'shared/lib/query/textQueryUtils';
 import {
     AndSearchClause,
     ISearchClause,
@@ -8,11 +8,14 @@ import {
     Phrase,
     QueryUpdate,
 } from 'shared/components/query/SearchClause';
+import { CheckboxFilterField } from 'shared/components/query/filteredSearch/CheckboxFilterField';
+import { QueryParser } from 'shared/lib/query/QueryParser';
 
 export type FilteredSearchDropdownFormProps = {
     query: ISearchClause[];
     filterConfig: CancerTreeSearchFilter[];
     onChange: (change: QueryUpdate) => void;
+    parser: QueryParser;
 };
 
 export const FilteredSearchDropdownForm: FunctionComponent<FilteredSearchDropdownFormProps> = props => {
@@ -27,8 +30,9 @@ export const FilteredSearchDropdownForm: FunctionComponent<FilteredSearchDropdow
                 return (
                     <FilterFormField
                         filter={filter}
-                        clauses={props.query}
+                        query={props.query}
                         onChange={props.onChange}
+                        parser={props.parser}
                     />
                 );
             })}
@@ -42,15 +46,6 @@ export const FilteredSearchDropdownForm: FunctionComponent<FilteredSearchDropdow
 export type FilterField = CheckboxFilterField | ListFilterField;
 
 /**
- * Options are checked when found in query
- */
-type CheckboxFilterField = {
-    input: typeof FilterCheckbox;
-    label: string;
-    options: string[];
-};
-
-/**
  * Options disappear when found in query:
  */
 type ListFilterField = {
@@ -58,10 +53,11 @@ type ListFilterField = {
     options: string[];
 };
 
-type FieldProps = {
+export type FieldProps = {
     filter: CancerTreeSearchFilter;
-    clauses: ISearchClause[];
+    query: ISearchClause[];
     onChange: (change: QueryUpdate) => void;
+    parser: QueryParser;
 };
 
 export const FilterFormField: FunctionComponent<FieldProps> = props => {
@@ -78,163 +74,18 @@ export const FilterFormField: FunctionComponent<FieldProps> = props => {
     );
 };
 
-export const FilterCheckbox: FunctionComponent<FieldProps> = props => {
-    const form = props.filter.form as CheckboxFilterField;
-    const prefix = props.filter.phrasePrefix || '';
-    let checkedPhrases: Phrase[] = [];
-    let uncheckedPhrases: Phrase[] = [];
-
-    const phrases = createPhrases(prefix, form.options);
-    const relevantClauses = props.clauses.filter(c =>
-        phrases.find(p => c.contains(p))
-    );
-    for (const phrase of phrases) {
-        const isChecked = isOptionChecked(phrase, relevantClauses);
-        if (isChecked) {
-            checkedPhrases.push(phrase);
-        } else {
-            uncheckedPhrases.push(phrase);
-        }
-    }
-
-    return (
-        <div className="filter-checkbox">
-            <span>{form.label}</span>
-            <div>
-                {phrases.map((option: Phrase) => {
-                    const id = `input-${option.phrase}`;
-                    let isChecked = checkedPhrases.includes(option);
-                    return (
-                        <div
-                            style={{
-                                display: 'inline-block',
-                                padding: '0 1em 0 0',
-                            }}
-                        >
-                            <input
-                                type="checkbox"
-                                id={id}
-                                value={option.phrase}
-                                checked={isChecked}
-                                onClick={() => {
-                                    isChecked = !isChecked;
-                                    updatePhrases(option, isChecked);
-                                    const update = createUpdate(
-                                        uncheckedPhrases,
-                                        checkedPhrases
-                                    );
-                                    props.onChange(update);
-                                }}
-                                style={{
-                                    display: 'inline-block',
-                                }}
-                            />
-                            <label
-                                htmlFor={id}
-                                style={{
-                                    display: 'inline-block',
-                                    padding: '0 0 0 0.2em',
-                                }}
-                            >
-                                {option.phrase}
-                            </label>
-                        </div>
-                    );
-                })}
-            </div>
-        </div>
-    );
-
-    function createPhrases(prefix: string, options: string[]): Phrase[] {
-        return options.map(option => {
-            const textRepresentation = `${prefix}:${option}`;
-            return {
-                phrase: option,
-                fields: props.filter.nodeFields,
-                textRepresentation,
-            };
-        });
-    }
-
-    function updatePhrases(phrase: Phrase, checked?: boolean) {
-        if (checked) {
-            checkedPhrases.push(phrase);
-            uncheckedPhrases = uncheckedPhrases.filter(as => as !== phrase);
-        } else {
-            uncheckedPhrases.push(phrase);
-            checkedPhrases = checkedPhrases.filter(as => as !== phrase);
-        }
-    }
-};
-
-/**
- * Determine if checkbox option is checked
- */
-function isOptionChecked(
-    phrase: Phrase,
-    relevantClauses: ISearchClause[]
-): boolean {
-    const clause = relevantClauses.find(c => c.contains(phrase));
-    if (clause && clause.isAnd()) {
-        return true;
-    }
-    if (clause && clause.isNot()) {
-        return false;
-    }
-    // option is checked when all others are not-clauses:
-    return (
-        relevantClauses.length === relevantClauses.filter(c => c.isNot()).length
-    );
-}
-
-/**
- * Create query update
- * while trying to keep query as short as possible
- */
-export function createUpdate(not: Phrase[], and: Phrase[]): QueryUpdate {
-    const toAdd: ISearchClause[] = [];
-    const toRemove: Phrase[] = [];
-
-    // if only and: remove all
-    if (!not.length) {
-        and.forEach(p => toRemove.push(p));
-    }
-    // if only not: create not
-    else if (!and.length) {
-        not.forEach(p => toAdd.push(new NotSearchClause(p)));
-    }
-    // if more and: create not, remove and
-    else if (and.length <= not.length) {
-        and.forEach(p => toAdd.push(new AndSearchClause([p])));
-        not.forEach(p => toRemove.push(p));
-    }
-    // if more not: create and, remove not
-    else {
-        and.forEach(p => toRemove.push(p));
-        not.forEach(p => toAdd.push(new NotSearchClause(p)));
-    }
-    return { toAdd, toRemove };
-}
-
 export const FilterList: FunctionComponent<FieldProps> = props => {
     const form = props.filter.form as ListFilterField;
-    const prefix = props.filter.phrasePrefix;
+
     return (
         <ul>
             {form.options.map(option => {
-                const textRepresentation = `${prefix}:${option}`;
-                const clause = new AndSearchClause([
-                    {
-                        phrase: option,
-                        fields: props.filter.nodeFields,
-                        textRepresentation,
-                    },
-                ]);
+                const update = props.parser.parseSearchQuery(option);
                 return (
                     <li className="menu-item">
                         <a
                             tabIndex={-1}
-                            onClick={() => props.onChange({ toAdd: [clause] })}
+                            onClick={() => props.onChange({ toAdd: update })}
                         >
                             {option}
                         </a>
